Moved all functions which make plots into a single plots.py file
