+++ FinalFigures/Gamma/figure2.py
@@ -1,61 +1,13 @@
-import matplotlib as mpl
 import matplotlib.pyplot as plt
 import matplotlib.pylab as pylab
-import matplotlib.colors as colors
-import src.Plotting.Mpl.AddCopyFig
 import numpy as np
-import lmfit as lm
-from itertools import chain
-
-from typing import Union
 
 import src.UsefulFunctions as U
-
-params = {'legend.fontsize': 'x-large',
-          'figure.figsize': (5, 4),
-          'axes.labelsize': 16,
-          'axes.titlesize': 20,
-          'xtick.labelsize': 16,
-          'xtick.direction': 'in',
-          'xtick.minor.visible': True,
-          'ytick.labelsize': 16,
-          'ytick.direction': 'in',
-          'ytick.minor.visible': True,
-          'text.usetex': False,
-          'font.family': 'sans-serif',
-          'font.sans-serif': ['Helvetica'],
-          }
-pylab.rcParams.update(params)
-
-
-def gamma_vs_coupling(ax: plt.Axes, coupling_gates: Union[list, np.ndarray],
-                      gammas: Union[list, np.ndarray]) -> plt.Axes:
-    """Adds Gamma vs Coupling plot to axes"""
-    # ax.set_title('Gamma/T vs Coupling Gate')
-    ax.set_xlabel('Coupling Gate /mV')
-    ax.set_ylabel('Gamma/KbT')
-
-    ax.plot(coupling_gates, gammas, marker='.', color='black')
-    return ax
-
-
-def amp_theta_vs_coupling(ax: plt.Axes, amp_coupling: Union[list, np.ndarray], amps: Union[list, np.ndarray],
-                          theta_coupling: Union[list, np.ndarray], thetas: Union[list, np.ndarray]) -> plt.Axes:
-    """Adds both amplitude vs coupling and theta vs coupling to axes"""
-    ax.set_title('Amplitude and dT vs Coupling Gate')
-    ax.set_xlabel('Coupling Gate /mV')
-    ax.set_ylabel('Amplitude /nA')
-
-    ax.plot(amp_coupling, amps)
-
-    ax2 = ax.twinx()
-    ax2.set_ylabel('dT /mV')
-    ax2.plot(theta_coupling, thetas)
-    return ax
+from FinalFigures.Gamma.plots import gamma_vs_coupling, amp_theta_vs_coupling, dndt_signal
 
 
 if __name__ == '__main__':
-    from src.DatObject.Make_Dat import get_dats, get_dat, DatHDF
+    from src.DatObject.Make_Dat import get_dats
 
     ####################################################
     # Data for gamma_vs_coupling
